Moved profile functions into its component
